requirements.txt
pyserial==3.5
hidapi==0.14.0.post4